Create view_and_pure.sol
This contract checks the difference between the view and pure. 
View ---> they are similar to the constant functions i.e used to read the data on the blockchain
Pure --->  they are view functions with the additional restriction that their value only depends on the function arguments. This means they cannot use SSTORE, SLOAD, cannot send or receive ether, cannot use msg or block and can only call other pure functions.